Melhorias no  Código
Adicionando classe ServiceException, correções nos calculos em VendaModel,  padronização valores em constants.

+++ Arquivo/src/br/com/ntconsult/arquivo/odata/ArquivoData.java
@@ -33,8 +33,14 @@ public class ArquivoData {
 	
 	public boolean deletarArquivo() {
 		
-		File file = new File(path);
-		return file.delete();
+		if( arquivoExiste() ) {
+		
+			File file = new File(path);
+			return file.delete();
+			
+		}
+		
+		return false;
 		
 	}
